Refactorings: move util.Preconditions into core and add exception package to core
