+++ src/main/java/pl/cleankod/exchange/entrypoint/ExceptionHandlerAdvice.java
@@ -3,7 +3,7 @@ package pl.cleankod.exchange.entrypoint;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import pl.cleankod.exchange.core.usecase.CurrencyConversionException;
+import pl.cleankod.exchange.core.exception.CurrencyConversionException;
 import pl.cleankod.exchange.entrypoint.model.ApiError;
 
 @ControllerAdvice
